Remove extra operations when generating pod sandbox configuration.
